Moved the tests to tests/Unit
