+++ tests/Unit/Archive/ArchiveListerTest.php
@@ -7,7 +7,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\SiteMigration\Test\Archive;
+namespace Piwik\Plugins\SiteMigration\tests\Unit\Archive;
 
 use Piwik\Plugins\SiteMigration\Migrator\Archive\ArchiveLister;
