+++ tests/Unit/DBHelperTest.php
@@ -7,7 +7,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\SiteMigration\Test;
+namespace Piwik\Plugins\SiteMigration\tests\Unit;
 
 use Piwik\Plugins\SiteMigration\Helper\DBHelper;
